daemon: move apiBaseSuite to its own api_base_test.go
APIBaseSuite will be used for a while both from daemon
an daemon_test, so opt to make some field exported
for now at least

also use AddCleanup more and correctly

some cleanups in api_users_test.go in preparation to make
live in daemon_test
